requirements.txt
# Python dependencies (install via pip)
openai>=1.0.0
pydantic>=2.0.0
pydub==0.25.1
click>=8.0.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
aiosqlite>=0.20.0
pytest-timeout>=2.2.0

# System dependencies (install via package manager)
# - dunst: Notification daemon (required for persistent notifications)
# - dmenu: Menu program (required for notification context menu actions)
# - dunstify: dunst's notify-send replacement (comes with dunst)
# Install on Arch Linux: sudo pacman -S dunst dmenu
# Install on Debian/Ubuntu: sudo apt-get install dunst dmenu